Configure onion service via torrc
Instead of configuring the Tor onion service in code using `torut` we
can configure it via the `torrc` file. This results in moving
the burden of configuration away from ping-pong and onto the user but
makes the code in ping-pong easier. However we provide a working
`torrc` so this burden is mitigated somewhat.