Enable listen_on to accept an onion address
Currently we pass the local address to `listen_on`, we need to pass an
onion address.

Add a hash map to map onion addresses to port numbers to use for the
local service. When creating a listener use the map to get a local
port number then listen on localhost:port.